pyproject.toml
[project]
name = "VQTEK-apply"
version = "0.1.0"
description = "VQTEK testwork"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pillow>=12.1.1",
    "requests>=2.32.5",
]
